fix for recurring events on dayname bug
This would occurr when someone selected "recurr on dayname" which would look for x dayname in month (ex: 3rd Thursday in a month). It was actually getting the week position of the day instead, which isn't always the same as the week position because a month doesn't always start on a Sunday.